feat: convert rucomfyui_generate_nodes to a library
Addresses #6: Convert Rust node generation to a library

Changes:
- Refactored main.rs to extract generation logic into lib.rs
- Created public API with `generate_nodes()` and `generate_nodes_with_config()`
- Added GenerateConfig struct for customization options
- Updated README with comprehensive library usage examples
- Binary interface remains unchanged for backward compatibility

This enables:
- Generating nodes for custom ComfyUI nodes
- Integration into build scripts and procedural macros
- Reuse of node generation logic in external projects

The library accepts arbitrary ObjectInfo parameters and generates
corresponding Rust code, enabling type-safe usage of any ComfyUI nodes.
